+++ R/Sql.R
@@ -490,7 +490,10 @@ querySql <- function(connection, sql, errorReportFile = file.path(getwd(), "erro
 #'                            there will be no progress bar, and no per-statement error messages. If the 
 #'                            database platform does not support batched updates the query is executed as 
 #'                            ordinarily.
-#' @param oracleTempSchema    A schema that can be used to create temp tables in when using Oracle or Impala.
+#' @param oracleTempSchema    DEPRECATED: use \code{tempEmulationSchema} instead.
+#' @param tempEmulationSchema Some database platforms like Oracle and Impala do not truly support temp tables. To
+#'                            emulate temp tables, provide a schema with write privileges where temp tables
+#'                            can be created.
 #' @param ...                 Parameters that will be used to render the SQL.
 #'
 #' @details
@@ -519,9 +522,14 @@ renderTranslateExecuteSql <- function(connection,
                                       errorReportFile = file.path(getwd(), "errorReportSql.txt"),
                                       runAsBatch = FALSE,
                                       oracleTempSchema = NULL,
+                                      tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
                                       ...) {
+  if (!is.null(oracleTempSchema) && oracleTempSchema != "") {
+    warning("The 'oracleTempSchema' argument is deprecated. Use 'tempEmulationSchema' instead.")
+    tempEmulationSchema <- oracleTempSchema
+  }
   sql <- SqlRender::render(sql, ...)
-  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
+  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = tempEmulationSchema)
   executeSql(connection = connection,
              sql = sql,
              profile = profile,
@@ -541,7 +549,10 @@ renderTranslateExecuteSql <- function(connection,
 #' @param errorReportFile      The file where an error report will be written if an error occurs. Defaults to
 #'                             'errorReportSql.txt' in the current working directory.
 #' @param snakeCaseToCamelCase If true, field names are assumed to use snake_case, and are converted to camelCase.  
-#' @param oracleTempSchema     A schema that can be used to create temp tables in when using Oracle or Impala.
+#' @param oracleTempSchema    DEPRECATED: use \code{tempEmulationSchema} instead.
+#' @param tempEmulationSchema Some database platforms like Oracle and Impala do not truly support temp tables. To
+#'                            emulate temp tables, provide a schema with write privileges where temp tables
+#'                            can be created.
 #' @param ...                  Parameters that will be used to render the SQL.
 #'
 #' @details
@@ -570,9 +581,14 @@ renderTranslateQuerySql <- function(connection,
                                     errorReportFile = file.path(getwd(), "errorReportSql.txt"), 
                                     snakeCaseToCamelCase = FALSE,
                                     oracleTempSchema = NULL,
+                                    tempEmulationSchema = getOption("sqlRenderTempEmulationSchema"),
                                     ...) {
+  if (!is.null(oracleTempSchema) && oracleTempSchema != "") {
+    warning("The 'oracleTempSchema' argument is deprecated. Use 'tempEmulationSchema' instead.")
+    tempEmulationSchema <- oracleTempSchema
+  }
   sql <- SqlRender::render(sql, ...)
-  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = oracleTempSchema)
+  sql <- SqlRender::translate(sql, targetDialect = connection@dbms, oracleTempSchema = tempEmulationSchema)
   return(querySql(connection = connection,
                   sql = sql,
                   errorReportFile = errorReportFile,